requirements.txt
Flask~=1.1.2
Flask-Migrate
connexion[swagger-ui]~=2.7.0
sqlalchemy~=1.4.6
pip~=20.2.4
wheel~=0.35.1
attrs~=20.3.0
Mako~=1.1.4
MarkupSafe~=1.1.1
PyYAML~=5.4.1
Jinja2~=2.11.3
Werkzeug~=1.0.1
itsdangerous~=1.1.0
click~=7.1.2
python-dateutil~=2.8.1
alembic~=1.5.8
certifi~=2020.12.5
chardet~=4.0.0
isodate~=0.6.0
six~=1.15.0
urllib3~=1.26.4
greenlet~=1.0.0
idna~=2.10
clickclick~=20.10.2
jsonschema~=3.2.0
requests~=2.25.1
inflection~=0.5.1
pyrsistent~=0.17.3
setuptools~=50.3.2
marshmallow~=3.11.1
PyJWT~=2.0.1